More improvements...
Switched the script to use Builder instead of the previous XML method.  This makes for simplier code and the exporting of minified files.  This was critical because KeePassX did not properly trim the information before importing.
Also added logic to parse newline escape characters to the actual newline character.  This allows the comments to be what they should be instead of gunked up with a bunch of "\n" bits everywhere.